+++ src/constants/index.ts
@@ -2,17 +2,18 @@
  * MDM centralised constants, referred as `@ukef/constants`.
  * Following constants are served:
  *
- * 1. Date
- * 2. Database name
- * 3. Auth strategy
- * 4. Products
- * 5. Customers
- * 6. Strings to redact
- * 7. Strings locations to redact
- * 8. Module geospatial
- * 9. GOV.UK notify email constants
- * 10. Companies module
- * 11. Companies House helper module
+ * Auth strategy
+ * Companies module
+ * Companies House helper module
+ * Customers
+ * Date
+ * Database name
+ * Deals
+ * GOV.UK notify email constants
+ * Module geospatial
+ * Products
+ * Strings to redact
+ * Strings paths to redact
  */
 
 export * from './auth.constant';
@@ -21,6 +22,7 @@ export * from './companies-house.constant';
 export * from './customers.constant';
 export * from './database.constant';
 export * from './date.constant';
+export * from './deals.constant';
 export * from './dun-and-bradstreet.constant';
 export * from './enums';
 export * from './geospatial.constant';