+++ src/modules/ods/dto/index.ts
@@ -1,0 +1,5 @@
+export * from './get-ods-customer-param.dto';
+export * from './get-ods-customer-response.dto';
+export * from './get-ods-deal-param.dto';
+export * from './get-ods-deal-response.dto';
+export * from './ods-payloads.dto';