+++ test/companies/get-company-by-registration-number.api-test.ts
@@ -3,7 +3,7 @@ import getCompanyCompaniesHouseOverseasCompanyResponse from '@ukef/helper-module
 import getCompanyResponse from '@ukef/modules/companies/examples/example-response-for-get-company-by-registration-number.json';
 import { IncorrectAuthArg, withClientAuthenticationTests } from '@ukef-test/common-tests/client-authentication-api-tests';
 import { Api } from '@ukef-test/support/api';
-import { ENVIRONMENT_VARIABLES, TIME_EXCEEDING_COMPANIES_HOUSE_TIMEOUT } from '@ukef-test/support/environment-variables';
+import { ENVIRONMENT_VARIABLES } from '@ukef-test/support/environment-variables';
 import { GetCompanyGenerator } from '@ukef-test/support/generator/get-company-generator';
 import { RandomValueGenerator } from '@ukef-test/support/generator/random-value-generator';
 import nock from 'nock';
@@ -40,7 +40,6 @@ describe('GET /companies?registrationNumber=', () => {
     nock.cleanAll();
   });
 
-  // MDM auth tests
   withClientAuthenticationTests({
     givenTheRequestWouldOtherwiseSucceed: () => {
       requestToGetCompanyByRegistrationNumber(companiesHousePath).reply(200, getCompanyCompaniesHouseResponse);
@@ -48,11 +47,13 @@ describe('GET /companies?registrationNumber=', () => {
     makeRequestWithoutAuth: (incorrectAuth?: IncorrectAuthArg) => api.getWithoutAuth(mdmPath, incorrectAuth?.headerName, incorrectAuth?.headerValue),
   });
 
-  it('returns a 200 response with the company if it is returned by the Companies House API', async () => {
+  it('returns a 200 response with the company when returned by Companies House API', async () => {
     requestToGetCompanyByRegistrationNumber(companiesHousePath).reply(200, getCompanyCompaniesHouseResponse);
 
+    // Act
     const { status, body } = await api.get(mdmPath);
 
+    // Assert
     expect(status).toBe(200);
     expect(body).toStrictEqual(getCompanyResponse);
   });
@@ -86,8 +87,10 @@ describe('GET /companies?registrationNumber=', () => {
   ])(
     'returns a 400 response with the correct validation errors if the registration number is $descriptionForTestName',
     async ({ registrationNumber, validationError }) => {
+      // Act
       const { status, body } = await api.get(getMdmPath(registrationNumber));
 
+      // Assert
       expect(status).toBe(400);
       expect(body).toMatchObject({
         error: 'Bad Request',
@@ -98,10 +101,13 @@ describe('GET /companies?registrationNumber=', () => {
   );
 
   it(`returns a 404 response if the Companies House API returns a 404 response status code`, async () => {
+    // Arrange
     requestToGetCompanyByRegistrationNumber(companiesHousePath).reply(404, getCompanyCompaniesHouseNotFoundResponse);
 
+    // Act
     const { status, body } = await api.get(mdmPath);
 
+    // Assert
     expect(status).toBe(404);
     expect(body).toStrictEqual({
       statusCode: 404,
@@ -110,12 +116,15 @@ describe('GET /companies?registrationNumber=', () => {
   });
 
   it('returns a 422 response if the Companies House API returns an overseas company', async () => {
+    // Arrange
     const registrationNumber = 'OE006930';
 
     requestToGetCompanyByRegistrationNumber(getCompaniesHousePath(registrationNumber)).reply(200, getCompanyCompaniesHouseOverseasCompanyResponse);
 
+    // Act
     const { status, body } = await api.get(getMdmPath(registrationNumber));
 
+    // Assert
     expect(status).toBe(422);
     expect(body).toStrictEqual({
       statusCode: 422,
@@ -124,10 +133,13 @@ describe('GET /companies?registrationNumber=', () => {
   });
 
   it(`returns a 500 response if the Companies House API returns a 400 response containing the error string 'Invalid Authorization header'`, async () => {
+    // Arrange
     requestToGetCompanyByRegistrationNumber(companiesHousePath).reply(400, getCompanyCompaniesHouseMalformedAuthorizationHeaderResponse);
 
+    // Act
     const { status, body } = await api.get(mdmPath);
 
+    // Assert
     expect(status).toBe(500);
     expect(body).toStrictEqual({
       statusCode: 500,
@@ -136,10 +148,13 @@ describe('GET /companies?registrationNumber=', () => {
   });
 
   it(`returns a 500 response if the Companies House API returns a 401 response containing the error string 'Invalid Authorization'`, async () => {
+    // Arrange
     requestToGetCompanyByRegistrationNumber(companiesHousePath).reply(401, getCompanyCompaniesHouseInvalidAuthorizationResponse);
 
+    // Act
     const { status, body } = await api.get(mdmPath);
 
+    // Assert
     expect(status).toBe(500);
     expect(body).toStrictEqual({
       statusCode: 500,
@@ -147,11 +162,14 @@ describe('GET /companies?registrationNumber=', () => {
     });
   });
 
-  it('returns a 500 response if the request to the Companies House API times out', async () => {
-    requestToGetCompanyByRegistrationNumber(companiesHousePath).delay(TIME_EXCEEDING_COMPANIES_HOUSE_TIMEOUT).reply(200, getCompanyCompaniesHouseResponse);
+  it('returns a 500 response if the request to the Companies House API returns a 500 status code', async () => {
+    // Arrange
+    requestToGetCompanyByRegistrationNumber(companiesHousePath).reply(500, getCompanyCompaniesHouseResponse);
 
+    // Act
     const { status, body } = await api.get(mdmPath);
 
+    // Assert
     expect(status).toBe(500);
     expect(body).toStrictEqual({
       statusCode: 500,
@@ -160,10 +178,13 @@ describe('GET /companies?registrationNumber=', () => {
   });
 
   it('returns a 500 response if the request to the Companies House API returns an unhandled error response', async () => {
+    // Arrange
     requestToGetCompanyByRegistrationNumber(companiesHousePath).reply(418, `I'm a teapot`);
 
+    // Act
     const { status, body } = await api.get(mdmPath);
 
+    // Assert
     expect(status).toBe(500);
     expect(body).toStrictEqual({
       statusCode: 500,