+++ test/ods/ods.deal.api-test.ts
@@ -1,12 +1,12 @@
 import { HttpStatus } from '@nestjs/common';
 import { Api } from '@ukef-test/support/api';
 
-describe('/ods/customers', () => {
+describe('/ods/deal', () => {
   let api: Api;
 
   const expectedResult = expect.objectContaining({
+    dealId: expect.any(String),
     name: expect.any(String),
-    urn: expect.any(String),
   });
 
   beforeAll(async () => {
@@ -17,10 +17,10 @@ describe('/ods/customers', () => {
     await api.destroy();
   });
 
-  describe('/:urn', () => {
-    it(`should return ${HttpStatus.OK} when the URN has a valid format and belongs to an existing customer`, async () => {
+  describe('/:id', () => {
+    it(`should return ${HttpStatus.OK} when the deal ID is a valid format and belongs to an existing deal`, async () => {
       // Act
-      const { status, body } = await api.get('/api/v1/ods/customers/00325182');
+      const { status, body } = await api.get('/api/v1/ods/deal/0000000436');
 
       // Assert
       expect(status).toBe(HttpStatus.OK);
@@ -28,44 +28,44 @@ describe('/ods/customers', () => {
       expect(body).toEqual(expectedResult);
     });
 
-    it(`should return ${HttpStatus.NOT_FOUND} when the URN has a valid format, but does not match an existing customer`, async () => {
+    it(`should return ${HttpStatus.NOT_FOUND} when the deal ID is a valid format, but does not match an existing deal`, async () => {
       // Act
-      const { status, body } = await api.get('/api/v1/ods/customers/99999999');
+      const { status, body } = await api.get('/api/v1/ods/deal/0000000001');
 
       // Assert
       expect(status).toBe(HttpStatus.NOT_FOUND);
 
       expect(body).toEqual({
         statusCode: HttpStatus.NOT_FOUND,
-        message: 'No customer found',
+        message: 'No deal found',
         error: 'Not Found',
       });
     });
 
-    it(`should return ${HttpStatus.BAD_REQUEST} when the URN is not the right length`, async () => {
+    it(`should return ${HttpStatus.BAD_REQUEST} when the deal ID is not the right length`, async () => {
       // Act
-      const { status, body } = await api.get('/api/v1/ods/customers/1234567');
+      const { status, body } = await api.get('/api/v1/ods/deal/1234567');
 
       // Assert
       expect(status).toBe(HttpStatus.BAD_REQUEST);
 
       expect(body).toEqual({
         statusCode: HttpStatus.BAD_REQUEST,
-        message: ['urn must match /^\\d{8}$/ regular expression'],
+        message: ['id must match /^00\\d{8}$/ regular expression'],
         error: 'Bad Request',
       });
     });
 
-    it(`should return ${HttpStatus.BAD_REQUEST} when the URN does not match the regex`, async () => {
+    it(`should return ${HttpStatus.BAD_REQUEST} when the deal ID does not match the regex`, async () => {
       // Act
-      const { status, body } = await api.get('/api/v1/ods/customers/abc');
+      const { status, body } = await api.get('/api/v1/ods/deal/abc');
 
       // Assert
       expect(status).toBe(HttpStatus.BAD_REQUEST);
 
       expect(body).toEqual({
         statusCode: HttpStatus.BAD_REQUEST,
-        message: ['urn must match /^\\d{8}$/ regular expression'],
+        message: ['id must match /^00\\d{8}$/ regular expression'],
         error: 'Bad Request',
       });
     });